Levels Project
First Level: More of a starting stage where I tested out a lot of phasing blocks and some teleporting features. For example when you landed in the water it would teleport you back to spawn. also used triggers in first level as test.

Second Level: Made it underwater themed kind of like a sequel/continuation of the first one. This one is more puzzle/scavenger hunt based and you need to activate switches to get to the next area. You start on the hand and proceed into the head. You are then let out to continue into the next area.

Third Level: The third level is a cloud level very much like the first level but easier. The switches are hidden in plain sight on the clouds and they are used to activate the platforms near the final cloud. The final cloud has a rest area and no teleporter, signifying the end. If you fall off the clouds, you get teleported back to the spawn cloud. There isn't much reason for making the cloud level other than it being a bit faster than the rest in terms of how quickly I can make the level.